--- src/App.test.tsx
@@ -1,10 +1,0 @@
-import { render, screen } from "@testing-library/react";
-import { App } from "./App";
-
-describe("Recipe", () => {
-  it('Should render "Recipes"', () => {
-    render(<App />);
-    const text = screen.getByText("Recipes");
-    expect(text).toBeDefined();
-  });
-});
